scripts: check_compliance: Check for kconfig variables declared
Add support for a single file as source of truth for environmental
variables declared for Kconfig in Cmake. It's coupled with a change in
the cmake build system where you declare the environmental variables for
Kconfig in a single text file. This will give you an error if the script
does not set the required variables and will list which variables are
missing.

Signed-off-by: Sigvart Hovland <sigvart.m@gmail.com>