check_compliance.py: Break Kconfig tests into several functions
This can be done in a neat way with the new
ComplianceTest.error()/skip()/add_failure() helpers.

Have a separate function for parsing the Kconfigs, and a separate
function for each test. Only parse once and reuse the kconfiglib.Kconfig
object between tests, to save a little bit of time.

Also allow tests to raise EndTest themselves wherever it's handy, and
document it.

Signed-off-by: Ulf Magnusson <Ulf.Magnusson@nordicsemi.no>