scripts: Fix risky uses of non-raw regex strings in Python scripts
Fixes pylint warnings like this one:

    scripts/check_compliance.py:326:0: W1401: Anomalous backslash in
    string: '\s'. String constant might be missing an r prefix.
    (anomalous-backslash-in-string)

The reason for this warning is that backslash escapes are interpreted in
non-raw (non-r-prefixed) strings. For example, '\a' and r'\a' are not
the same string (first one has a single ASCII bell character, second one
has two characters).

It just happens that there's no \s (or \., or \/) escape for example,
and '\s' turns into two characters (as needed for a regex). It's risky
to rely on stuff like that regexes though. Best to make them raw strings
unless they're super trivial.

Also note that '\s' and '\\s' turn into the same string.

Another tip: A literal ' can be put into a string with "blah'blah"
instead of 'blah\'blah'.

Piggyback removal of a redundant semicolon.

zephyrproject-rtos/ci-tools#37 (just to link)

Signed-off-by: Ulf Magnusson <Ulf.Magnusson@nordicsemi.no>