requirements: require urllib3>=1.24.2
CVE-2019-11324 More information

The urllib3 library before 1.24.2 for Python mishandles certain cases
where the desired set of CA certificates is different from the OS store
of CA certificates, which results in SSL connections succeeding in
situations where a verification failure is the correct outcome. This is
related to use of the ssl_context, ca_certs, or ca_certs_dir argument.

Signed-off-by: Anas Nashif <anas.nashif@intel.com>