check_compliance.py: run gitlint in GIT_TOP
gitlint is very "flexible" - and confusing: it can run from any
sub-directory but looks for .gitlint configuration only in the current
directory. For instance running:

  check_compliance.py -c 4fc250489d~..

... in this repo reports a slightly different set of issues depending on
what the current sub-directory is.

Zephyr is not interested in per-directory rules right now; always run
gitlint in GIT_TOP where the single .gitlint is.

Signed-off-by: Marc Herbert <marc.herbert@intel.com>