codeowners: temporarily disable codeowner failures
We are reporting issues after the fact, i.e. a file a is removed in one
PR and entry in CODEOWNER is left, no errors. We merge the PR and all
seems fine until the next PR comes in, we then start reporting issues
with CODEOWNERS which is not related to the current PR.

Disable for now until we have a proper fix.

Signed-off-by: Anas Nashif <anas.nashif@intel.com>