check_compliance.py: Fix dtlib/edtlib tests and require Python 3.6
DeviceTreeCheck needs to catch SystemExit (raised by sys.exit()) instead
of Exception when running the dtlib and edtlib test suites, because
that's how they report failures (maybe it could be tweaked later).
SystemExit is not a subclass of Exception.

The dtlib/edtlib test suites also rely on dictionaries preserving
insertion order. They don't before Python 3.6. The code could be updated
to use collections.OrderedDict to work around it, but I'm not sure it's
worth it with Python 3.6 starting to get old (released December 2016).
Skip the test for Python 3.5 and earlier for now.

I'll make sure to run the test suites manually for PRs in the meantime,
until CI is updated to use Python 3.6.

Signed-off-by: Ulf Magnusson <Ulf.Magnusson@nordicsemi.no>