check_compliance.py: Whitelist example symbols from the script itself
These are mentioned in the script itself. They need to be whitelisted
before ci-tools can be merged back into the main Zephyr repository,
because that makes the Kconfig checks run on check_compliance.py itself.

Use CONFIG_FOO as an example instead of CONFIG_BAZ to fix one of the
references. CONFIG_FOO is already whitelisted.

Signed-off-by: Ulf Magnusson <Ulf.Magnusson@nordicsemi.no>