check_compliance.py: Make license check informational
Since this check is often a manual-inspection type thing, turn it into
an informational message instead of a failure. That might avoid people
needlessly changing stuff just to please the check and get a green
checkmark, or people holding off on their review because they think
there's problems with the code.

Also encourage a compact style in the message.

Signed-off-by: Ulf Magnusson <Ulf.Magnusson@nordicsemi.no>