check_compliance.py: Detect refs to undef. symbols in samples and tests
Extend the check for references to undefined Kconfig symbols to also
detect undefined symbols in samples and tests.

Samples and tests were skipped due to using separate Kconfig trees,
which hid the symbols defined in them. Work around it by grepping for
Kconfig symbol definitions in them instead.

Keep properly parsing the main Kconfig tree, as it's needed to see
symbol names that are stitched together with the Kconfig preprocessor.

Signed-off-by: Ulf Magnusson <Ulf.Magnusson@nordicsemi.no>