check_compliance.py: Fix running Kconfig/pylint checks outside top dir
The PyLint and KconfigCheck tests are broken when not run from the
top-level repository directory. Fix it by making the path to pylintrc
absolute and explicitly running the 'git grep' for Kconfig definitions
in samples/ and tests/ from the top-level Zephyr directory.

The root cause was a mismatch between paths and process working
directories.

Signed-off-by: Ulf Magnusson <Ulf.Magnusson@nordicsemi.no>