check_compliance.py: Add helper for formatting commands as strings
The " ".join(shlex.quote(word) for word in cmd) pattern was starting to
get copied in a bunch of places. Add a helper for it.

Piggyback a small improvement to the Pylint test:

I discovered earlier that FileNotFoundError is also generated when the
cwd= passed to subprocess.Popen() doesn't exists, giving a really
confusing error in that case. Just remove the FileNotFoundError check.
It's a subclass of OSError, so it'll still be caught.

Signed-off-by: Ulf Magnusson <Ulf.Magnusson@nordicsemi.no>